+++ src/main/java/com/myfactory/SBootWebProject/repository/empleado/EmpleadoJPABaseRepository.java
@@ -3,7 +3,7 @@ package com.myfactory.SBootWebProject.repository.empleado;
 import org.springframework.data.repository.CrudRepository;
 import com.myfactory.SBootWebProject.model.Empleado;
 
-public interface EmpleadoJPABaseRepository extends CrudRepository<Empleado, Long>{ 
+public interface EmpleadoJPABaseRepository extends CrudRepository<Empleado, Integer>{ 
 
 	  
 }
